server/pyproject.toml
[tool.poetry]
name = "megacad"
version = "0.1.0"
description = ""
authors = ["Entropy <70720874+NavinKumarMNK@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.8"
django = "4.2.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
whitenoise = "^6.6.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
aiohttp = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
